# Parallelize Placeholder loading for an LCP improvement