--- a/libs/blocks/chart/chart.js
+++ b/libs/blocks/chart/chart.js
@@ -1,4 +1,4 @@
-import { loadScript, getConfig, createTag } from '../../utils/utils.js';
+import { loadScript, getConfig, createTag, customFetch } from '../../utils/utils.js';
 import {
   throttle,
   parseValue,
@@ -122,7 +122,6 @@ export function processMarkData(series, xUnit) {
 }
 
 export async function fetchData(link) {
-  const { customFetch } = await import('../../utils/helpers.js');
   const resp = await customFetch({ resource: link.href.toLowerCase(), withCacheRules: true })
     .catch(() => ({}));