--- a/libs/features/placeholders.js
+++ b/libs/features/placeholders.js
@@ -1,3 +1,5 @@
+import { customFetch, getConfig } from '../utils/utils.js';
+
 const fetchedPlaceholders = {};
 window.mph = {};
 
@@ -7,26 +9,23 @@ const getPlaceholdersPath = (config, sheet) => {
   return `${path}${query}`;
 };
 
-const fetchPlaceholders = async (config, sheet) => {
-  const placeholdersPath = getPlaceholdersPath(config, sheet);
-  const { customFetch } = await import('../utils/helpers.js');
-
-  fetchedPlaceholders[placeholdersPath] = fetchedPlaceholders[placeholdersPath]
-    // eslint-disable-next-line no-async-promise-executor
-    || new Promise(async (resolve) => {
-      const resp = await customFetch({ resource: placeholdersPath, withCacheRules: true })
-        .catch(() => ({}));
-      const json = resp.ok ? await resp.json() : { data: [] };
-      if (json.data.length === 0) { resolve({}); return; }
-      const placeholders = {};
-      json.data.forEach((item) => {
-        placeholders[item.key] = item.value;
-        window.mph[item.key] = item.value;
-      });
-      resolve(placeholders);
+const fetchPlaceholders = async ({ config, sheet, placeholderRequest, placeholderPath }) => {
+  const path = placeholderPath || getPlaceholdersPath(config, sheet);
+  // eslint-disable-next-line no-async-promise-executor
+  fetchedPlaceholders[path] = fetchedPlaceholders[path] || new Promise(async (resolve) => {
+    const resp = await placeholderRequest || await customFetch(
+      { resource: path, withCacheRules: true },
+    ).catch(() => ({}));
+    const json = resp.ok ? await resp.json() : { data: [] };
+    if (json.data.length === 0) { resolve({}); return; }
+    const placeholders = {};
+    json.data.forEach((item) => {
+      window.mph[item.key] = item.value;
+      placeholders[item.key] = item.value;
     });
-
-  return fetchedPlaceholders[placeholdersPath];
+    resolve(placeholders);
+  });
+  return fetchedPlaceholders[path];
 };
 
 function keyToStr(key) {
@@ -60,7 +59,7 @@ async function getPlaceholder(key, config, sheet) {
       },
     };
 
-    const defaultPlaceholders = await fetchPlaceholders(defaultConfig, sheet)
+    const defaultPlaceholders = await fetchPlaceholders({ config: defaultConfig, sheet })
       .catch(() => ({}));
     defaultFetched = true;
     return defaultPlaceholders;
@@ -68,7 +67,7 @@ async function getPlaceholder(key, config, sheet) {
 
   if (config.placeholders?.[key]) return config.placeholders[key];
 
-  const placeholders = await fetchPlaceholders(config, sheet).catch(async () => {
+  const placeholders = await fetchPlaceholders({ config, sheet }).catch(async () => {
     const defaultPlaceholders = await getDefaultPlaceholders();
     return defaultPlaceholders;
   });
@@ -102,18 +101,38 @@ export async function replaceKeyArray(keys, config, sheet = 'default') {
   return placeholders;
 }
 
-export async function replaceText(text, config, regex = /{{(.*?)}}|%7B%7B(.*?)%7D%7D/g, sheet = 'default') {
+export async function replaceText(
+  text,
+  config,
+  regex = /{{(.*?)}}|%7B%7B(.*?)%7D%7D/g,
+  sheet = 'default',
+) {
   if (typeof text !== 'string' || !text.length) return '';
 
   const matches = [...text.matchAll(new RegExp(regex))];
   if (!matches.length) {
     return text;
   }
-  const keys = Array.from(matches, (match) => (match[1] || match[2]));
+  const keys = Array.from(matches, (match) => match[1] || match[2]);
   const placeholders = await replaceKeyArray(keys, config, sheet);
   // The .shift method is very slow, thus using normal iterator
   let i = 0;
   // eslint-disable-next-line no-plusplus
   const finalText = text.replaceAll(regex, () => placeholders[i++]);
   return finalText;
+}
+
+export async function decoratePlaceholderArea({
+  placeholderPath,
+  placeholderRequest,
+  nodes,
+}) {
+  if (!nodes.length) return;
+  const config = getConfig();
+  await fetchPlaceholders({ placeholderPath, config, placeholderRequest });
+  const replaceNodes = nodes.map(async (textNode) => {
+    textNode.nodeValue = await replaceText(textNode.nodeValue, config);
+    textNode.nodeValue = textNode.nodeValue.replace(/&nbsp;/g, '\u00A0');
+  });
+  await Promise.all(replaceNodes);
 }