--- a/libs/utils/helpers.js
+++ b/libs/utils/helpers.js
@@ -19,12 +19,3 @@ export function updateLinkWithLangRoot(link) {
     return link;
   }
 }
-
-export async function customFetch({ resource, withCacheRules }) {
-  const options = {};
-  if (withCacheRules) {
-    const params = new URLSearchParams(window.location.search);
-    options.cache = params.get('cache') === 'off' ? 'reload' : 'default';
-  }
-  return fetch(resource, options);
-}