--- a/test/blocks/caas/mocks/utils.js
+++ b/test/blocks/caas/mocks/utils.js
@@ -12,6 +12,8 @@ export const utf8ToB64 = (str) => window.btoa(unescape(encodeURIComponent(str)))
 
 export const b64ToUtf8 = (str) => decodeURIComponent(escape(window.atob(str)));
 
+export const customFetch = stub();
+
 export function getMetadata(name, doc = document) {
   const attr = name && name.includes(':') ? 'property' : 'name';
   const meta = doc.head.querySelector(`meta[${attr}="${name}"]`);