--- a/test/blocks/marketo/mocks/marketo-utils.js
+++ b/test/blocks/marketo/mocks/marketo-utils.js
@@ -67,3 +67,5 @@ export const localizeLink = (href) => href;
 export const loadLink = stub().returns(new Promise((resolve) => {
   resolve();
 }));
+
+export const customFetch = stub();