--- a/test/blocks/merch/mocks/embed-utils.js
+++ b/test/blocks/merch/mocks/embed-utils.js
@@ -27,6 +27,8 @@ export const getConfig = () => config;
 
 export const setConfig = (c) => { config = c; };
 
+export const customFetch = stub();
+
 export const loadArea = stub();
 
 export const loadScript = stub();