### Description

- Load placeholders for each section and not for the whole document. No placeholders in LCP? No placeholders loading
- Parallelize loading placeholders together with their JS
- Move `customFetch` into `utils.js` for faster LCP loading times for certain blocks & placeholders

There are substantial gains in this PR for high ping (mobile) & a bit less for low-ping (desktop) - but in both scenarios we should save over 100ms on the LCP

On my desktop I saw a 400 MS+ LCP improvement on an uncached hlx.page
Before: `195(placeholders.js) + 200 (helpers.js) + 300 (placeholders.json)` = 695 MS request time
After: 300ms (placeholders.js + placeholders.json)


**1. Sequential load**
Currently we load placeholders.js, then we load helpers.js to use customFetch (fetch without cache headers) then we load placeholders.json .... On a fast 3G network aka "high ping environment" that's 3x400ms per request
![1  sequential load](https://github.com/user-attachments/assets/90934276-7c2b-487a-bf03-122c13347728)

**2. Parallelising the loading**
If we shift the loading sequence to load `placeholders.js` + `placeholders.json` together, we can save one 400ms request
![2  parallel load -30-600ms](https://github.com/user-attachments/assets/6fe529c0-b749-41ee-99b0-57dc4657b0b9)

**3. Removing the customFetch dependency**
We are waiting another 400ms from another file for 7 lines of code. We can move those into `utils.js` to save on this precious pre-LCP time on mobile.
![3  remove helpers dependency another -30-600ms](https://github.com/user-attachments/assets/74507126-c4cf-496e-bcb9-104d49555cb2)

Resolves: [MWPW-156840](https://jira.corp.adobe.com/browse/MWPW-156840)

**Test URLs:**
- Before: https://main--milo--adobecom.hlx.page/drafts/osahin/gnav-1?martech=off
- After: https://placeholder-lcp-improvements--milo--mokimo.hlx.page/drafts/osahin/gnav-1?martech=off