@mokimo 2024-08-16
Still resolving a bug in CC that fetches placeholders a few too many times... Still want to open this already (mainly to not lose the precious description I already done)