@mokimo 2024-08-16
I went a step further to ensure we build the placeholders on a per-section-basis AND only if we have them.
If section 5 has placeholders, there is no reason we eagerly load them before the LCP.
![image](https://github.com/user-attachments/assets/77d62681-294e-4feb-85b2-5a274d193d12)