@mokimo 2024-08-21
CC test links
- Before: https://main--cc--adobecom.hlx.live/de/products/photoshop?martech=off
- After: https://main--cc--adobecom.hlx.live/de/products/photoshop?milolibs=placeholder-lcp-improvements--milo--mokimo&martech=off